Get iTrust deploying in digital ocean

+++ commands/deploy.js
@@ -33,11 +33,13 @@ exports.handler = async argv => {
 
 async function run(inventory, application){
 
-    result = sshSync(`cp /bakerx/.ssh/* /home/vagrant/.ssh/`, 'vagrant@192.168.33.20');
-    if( result.error ) { process.exit( result.status ); }
-
-    result = sshSync(`chmod 600 /home/vagrant/.ssh/digitalocean_rsa`, 'vagrant@192.168.33.20');
-    if( result.error ) { process.exit( result.status ); }
+    if (fs.existsSync('.ssh/digitalocean_rsa')) {
+        result = sshSync(`cp /bakerx/.ssh/* /home/vagrant/.ssh/`, 'vagrant@192.168.33.20');
+        if( result.error ) { process.exit( result.status ); }
+    
+        result = sshSync(`chmod 600 /home/vagrant/.ssh/digitalocean_rsa`, 'vagrant@192.168.33.20');
+        if( result.error ) { process.exit( result.status ); }
+    }
 
     let filePath = '';
